Refactored how scenes are handled
