+++ src/scene/lsystem/bounding_box.rs
